moved selectSkillsByUsedId method to skill manager
