+++ src/Model/OfferingManager.php
@@ -6,7 +6,6 @@ class OfferingManager extends AbstractManager
 {
     public const TABLE_PERSON = 'person';
     public const TABLE = 'offering';
-    public const TABLE_PERSON_SKILL = 'user_skill';
 
     public function insert(array $offering): int
     {
@@ -47,20 +46,6 @@ class OfferingManager extends AbstractManager
         $statement->execute();
 
         return $statement->fetch();
-    }
-
-    public function selectSkillsByUserId(int $userId): array
-    {
-        $statement = $this->pdo->prepare("SELECT s.name, us.person_id 
-        FROM " . static::TABLE_PERSON_SKILL . " AS us
-        JOIN " . self::TABLE_PERSON . " AS p ON us.person_id=p.id 
-        LEFT JOIN " . SkillManager::TABLE . " AS s ON s.id=us.skill_id
-        HAVING us.person_id=:userid
-        ORDER BY s.name ASC;");
-        $statement->bindValue('userid', $userId, \PDO::PARAM_INT);
-        $statement->execute();
-
-        return $statement->fetchAll();
     }
