+++ src/Model/SkillManager.php
@@ -7,6 +7,8 @@ use PDO;
 class SkillManager extends AbstractManager
 {
     public const TABLE = 'skill';
+    public const TABLE_PERSON = 'person';
+    public const TABLE_PERSON_SKILL = 'user_skill';
 
     /** Insert skill off a person in database */
     public function assignSkill(int $userId, array $skillIds): void
@@ -18,5 +20,19 @@ class SkillManager extends AbstractManager
             $statement->bindValue('userId', $userId, PDO::PARAM_INT);
             $statement->execute();
         }
+    }
+
+    public function selectSkillsByUserId(int $userId): array
+    {
+        $statement = $this->pdo->prepare("SELECT s.name, us.person_id 
+        FROM " . static::TABLE_PERSON_SKILL . " AS us
+        JOIN " . self::TABLE_PERSON . " AS p ON us.person_id=p.id 
+        LEFT JOIN " . self::TABLE . " AS s ON s.id=us.skill_id
+        HAVING us.person_id=:userid
+        ORDER BY s.name ASC;");
+        $statement->bindValue('userid', $userId, \PDO::PARAM_INT);
+        $statement->execute();
+
+        return $statement->fetchAll();
     }
 }
